Renames rigidbody file
